I have made a door
have made a model, a script, animations.
the door has options of Automatic opening , opening with Key (not
keyboard key but lock key [ can add a finding dimension in the game ],
if the door is hackable [ I just overdid it with this one i guess] and
OF COURSE the door can be destroyed.
Destroying the doors doesnt look beautiful though.